requirements.txt
google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
functions-framework
google-cloud-storage
google-cloud-logging	
google-cloud-aiplatform
flask
google-genai
google-cloud-aiplatform
google-cloud-storage
google-cloud-logging
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai